refactor : payment 엔티티 설정 시 필요한 reservation 엔티티 수정
